Remove ke::IRunnable.
This was a poor API from the get-go, at the very least it should have
been a subclass of the Thread type. When using the C++11 lambda version of
Thread, on Posix systems, the "IRunnable" was uninitialized and could
crash.

We might as well take this opportunity to remove IRunnable since it's no
longer needed.