Remove AString and replace uses with std::string.
It's difficult to mix AString with third-party C++ code that uses
std::string. The only advantage AString provides is being able to
transfer ownership of the underlying buffer, but that is not enough to
warrant maintaining this half-implemented class.

Also, fix a bug where Vector did not call the destructor of moved
objects. This was noticed when debugging an issue with a vector of
strings.