Fix StopOption tests
From the StopOption class comment:

This is a ToggleOption that will stop argument processing and force
parse_args to return true. It is the responsibility of the implementor
to check all StopOptions before any other options, as a true return
from parse_args() may have only filled one of them.