Fix argparser StopOption still requiring positional arguments (#115)
* Fix argparser StopOption still requiring positional arguments
* Fix StopOption tests

Even after encountering a StopOption, the parser would check if there are enough positional parameters leading to a different error message than what the implementor wanted for the StopOption.

From the StopOption class comment:

This is a ToggleOption that will stop argument processing and force
parse_args to return true. It is the responsibility of the implementor
to check all StopOptions before any other options, as a true return
from parse_args() may have only filled one of them.